// src/protos/ciot/proto/v2/iface.ts
export interface IfaceInfo {
    /**
     * @generated from protobuf field: uint32 id = 1
     */
    id: number; // Interface ID.
    /**
     * @generated from protobuf field: Ciot.IfaceType type = 2
     */
    type: IfaceType; // Interface type.
}
// ...
export enum IfaceType {
    /**
     * Undefined interface type.
     *
     * @generated from protobuf enum value: IFACE_TYPE_UNDEFINED = 0;
     */
    UNDEFINED = 0,
    /**
     * Custom interface type.
     *
     * @generated from protobuf enum value: IFACE_TYPE_CUSTOM = 1;
     */
    CUSTOM = 1,
    /**
     * CIOT interface type.
     *
     * @generated from protobuf enum value: IFACE_TYPE_CIOT = 2;
     */
    CIOT = 2,
    /**
     * Storage interface type.
     *
     * @generated from protobuf enum value: IFACE_TYPE_STORAGE = 3;
     */
    STORAGE = 3,
    /**
     * System interface type.
     *
     * @generated from protobuf enum value: IFACE_TYPE_SYS = 4;
     */
    SYS = 4,
    /**
     * UART interface type.
     *
     * @generated from protobuf enum value: IFACE_TYPE_UART = 5;
     */
    UART = 5,
    /**
     * USB interface type.
     *
     * @generated from protobuf enum value: IFACE_TYPE_USB = 6;
     */
    USB = 6,
    /**
     * TCP interface type.
     *
     * @generated from protobuf enum value: IFACE_TYPE_TCP = 7;
     */
    TCP = 7,
    /**
     * Ethernet interface type.
     *
     * @generated from protobuf enum value: IFACE_TYPE_ETH = 8;
     */
    ETH = 8,
    /**
     * WiFi interface type.
     *
     * @generated from protobuf enum value: IFACE_TYPE_WIFI = 9;
     */
    WIFI = 9,
    /**
     * BLE interface type.
     *
     * @generated from protobuf enum value: IFACE_TYPE_BLE = 10;
     */
    BLE = 10,
    /**
     * BLE SCN interface type.
     *
     * @generated from protobuf enum value: IFACE_TYPE_BLE_SCN = 11;
     */
    BLE_SCN = 11,
    /**
     * BLE ADV interface type.
     *
     * @generated from protobuf enum value: IFACE_TYPE_BLE_ADV = 12;
     */
    BLE_ADV = 12,
    /**
     * GPIO interface type.
     *
     * @generated from protobuf enum value: IFACE_TYPE_GPIO = 13;
     */
    GPIO = 13,
    /**
     * NTP interface type.
     *
     * @generated from protobuf enum value: IFACE_TYPE_NTP = 14;
     */
    NTP = 14,
    /**
     * OTA interface type.
     *
     * @generated from protobuf enum value: IFACE_TYPE_OTA = 15;
     */
    OTA = 15,
    /**
     * DFU interface type.
     *
     * @generated from protobuf enum value: IFACE_TYPE_DFU = 16;
     */
    DFU = 16,
    /**
     * HTTP Client interface type.
     *
     * @generated from protobuf enum value: IFACE_TYPE_HTTP_CLIENT = 17;
     */
    HTTP_CLIENT = 17,
    /**
     * HTTP Server interface type.
     *
     * @generated from protobuf enum value: IFACE_TYPE_HTTP_SERVER = 18;
     */
    HTTP_SERVER = 18,
    /**
     * MQTT interface type.
     *
     * @generated from protobuf enum value: IFACE_TYPE_MQTT_CLIENT = 19;
     */
    MQTT_CLIENT = 19,
    /**
     * Socket interface type.
     *
     * @generated from protobuf enum value: IFACE_TYPE_SOCKET = 20;
     */
    SOCKET = 20,
    /**
     * Bridge interface type.
     *
     * @generated from protobuf enum value: IFACE_TYPE_BRIDGE = 21;
     */
    BRIDGE = 21,
    /**
     * IOTA client interface type.
     *
     * @generated from protobuf enum value: IFACE_TYPE_IOTA_CLIENT = 22;
     */
    IOTA_CLIENT = 22,
    /**
     * IOTA server interface type.
     *
     * @generated from protobuf enum value: IFACE_TYPE_IOTA_SERVER = 23;
     */
    IOTA_SERVER = 23,
    /**
     * LOG interface type.
     *
     * @generated from protobuf enum value: IFACE_TYPE_LOG = 24;
     */
    LOG = 24,
    /**
     * Modbus client interface type.
     *
     * @generated from protobuf enum value: IFACE_TYPE_MBUS_CLIENT = 25;
     */
    MBUS_CLIENT = 25,
    /**
     * Modbus server interface type.
     *
     * @generated from protobuf enum value: IFACE_TYPE_MBUS_SERVER = 26;
     */
    MBUS_SERVER = 26,
    /**
     * PWM interface type.
     *
     * @generated from protobuf enum value: IFACE_TYPE_PWM = 27;
     */
    PWM = 27
}
// ...
export enum DataType {
    /**
     * Unknown data type.
     *
     * @generated from protobuf enum value: DATA_TYPE_UNKNOWN = 0;
     */
    UNKNOWN = 0,
    /**
     * Stop data type.
     *
     * @generated from protobuf enum value: DATA_TYPE_STOP = 1;
     */
    STOP = 1,
    /**
     * Configuration data type.
     *
     * @generated from protobuf enum value: DATA_TYPE_CONFIG = 2;
     */
    CONFIG = 2,
    /**
     * Status data type.
     *
     * @generated from protobuf enum value: DATA_TYPE_STATUS = 3;
     */
    STATUS = 3,
    /**
     * Request data type.
     *
     * @generated from protobuf enum value: DATA_TYPE_REQUEST = 4;
     */
    REQUEST = 4,
    /**
     * Information data type.
     *
     * @generated from protobuf enum value: DATA_TYPE_INFO = 5;
     */
    INFO = 5
}
// ...
class IfaceInfo$Type extends MessageType<IfaceInfo> {
    constructor() {
        super("Ciot.IfaceInfo", [
            { no: 1, name: "id", kind: "scalar", T: 13 /*ScalarType.UINT32*/ },
            { no: 2, name: "type", kind: "enum", T: () => ["Ciot.IfaceType", IfaceType, "IFACE_TYPE_"] }
        ]);
    }
    create(value?: PartialMessage<IfaceInfo>): IfaceInfo {
        const message = globalThis.Object.create((this.messagePrototype!));
        message.id = 0;
        message.type = 0;
        if (value !== undefined)
            reflectionMergePartial<IfaceInfo>(this, message, value);
        return message;
    }
    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: IfaceInfo): IfaceInfo {
        let message = target ?? this.create(), end = reader.pos + length;
        while (reader.pos < end) {
            let [fieldNo, wireType] = reader.tag();
            switch (fieldNo) {
                case /* uint32 id */ 1:
                    message.id = reader.uint32();
                    break;
                case /* Ciot.IfaceType type */ 2:
                    message.type = reader.int32();
                    break;
                default:
                    let u = options.readUnknownField;
                    if (u === "throw")
                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
                    let d = reader.skip(wireType);
                    if (u !== false)
                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
            }
        }
        return message;
    }
    internalBinaryWrite(message: IfaceInfo, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
        /* uint32 id = 1; */
        if (message.id !== 0)
            writer.tag(1, WireType.Varint).uint32(message.id);
        /* Ciot.IfaceType type = 2; */
        if (message.type !== 0)
            writer.tag(2, WireType.Varint).int32(message.type);
        let u = options.writeUnknownFields;
        if (u !== false)
            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
        return writer;
    }
}
/**
 * @generated MessageType for protobuf message Ciot.IfaceInfo
 */
export const IfaceInfo = new IfaceInfo$Type();